Fix some comments.
Get rid of null-pointer tests for t2i in the "not null" branch of an
earlier test whether it's null, as those tests are redundant.

Use a switch statement to check the subcommand for Trans2.

If t2i->info_level is -1, it means we don't know the info level, for
whatever reason (e.g., the request was cut short by the snapshot length
before the info level).  Report it as such.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@37183 f5534014-38df-0310-8fa8-9805f1628bb7