From Roland Knall via bug 5977:
Handling of SercosIII if the plugin can not be loaded.

Changed the behaviour - if the SercosIII plugin is not available, the
openSAFETY dissector registers itself for ETHERTYPE_SERCOS. SercosIII
packages can not be dissected anyway in such a case and currently no
other dissector will handle those packages.

The following steps will be taken to display the situation to the user:

The message "SercosIII dissector not available, openSAFETY/SercosIII
native dissection." will be added to the tree. Additionally a similar
warning will be presented on the console during startup of wireshark.

All openSAFETY packages will be dissected, displaying the normal
openSAFETY information.

This will be more transparent for the customer.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@37533 f5534014-38df-0310-8fa8-9805f1628bb7