Add a URL for the specifications.  (No, the Wireshark source code is not a specification.)
Put V1 support back, in case there are captures out there with V1
packets and somebody wants to read them.

We don't need two identical tests against the minimum length.

For an unknown version, just put in a text item for the data, don't
append a note to the length field (which is the wrong field to append to
in any case).

Make routines not used outside this file static.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@37912 f5534014-38df-0310-8fa8-9805f1628bb7