From Mike Morrin:  GSM ENHANCED MEASUREMENT REPORT PDUs were not dissected when present as L3_INFO in RSL MEAS_RES PDUs.
It seems that the RSL L3_INFO needs to be handled by a different dissector
depending on whether it contains a DTAP, SACCH or CCCH PDU, which fortunately
can be deduced from the RSL PDU type.  packet-rsl.c is updated to implement
this.

In packet-gsm_a_rr.c the dissection of PDUs with RR Short PD format is
improved, and also some items are renamed to make clearer the difference
between SACCH PDUs (which cna be normal or Short PD format) and RR Short PD
format PDUs (which can occur on SACCH, CCCH, or DCH).

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6699

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@40357 f5534014-38df-0310-8fa8-9805f1628bb7