With the WTAP_ERR_UNSUPPORTED error, Wiretap supplies a string giving the details of what in particular is unsupported; report it in TShark and Wireshark.
Handle WTAP_ERR_RANDOM_OPEN_PIPE in TShark.

Handle WTAP_ERR_COMPRESSION_NOT_SUPPORTED in TShark, and have its error
message in Wireshark not speak of gzip, in case we support compressed
output in other formats in the future.

If we see a second section header block in a pcap-NG file, don't report
it as "the file is corrupted", report it as "the file uses a feature we
don't support", as that's the case - and don't free up the interface
data array, as the file remains open, and Wireshark might still try to
access the packets we were able to read.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@41041 f5534014-38df-0310-8fa8-9805f1628bb7