Add an explicit capture_cb_capture_failed indication for the case where we weren't even able to start a capture, rather than delivering a fake "capture start" indication and relying on a later "capture file closed" indication - for a capture that was never opened in the first place - to handle GUI cleanups.
Don't deliver any GUI indications in cf_close() if we didn't have a
capture file open in the first place.

Clear the status bar and welcome header if that indication is delivered.

If we start a capture from the command line with the -k flag, don't show
the captured packet information unless the capture actually starts.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@41521 f5534014-38df-0310-8fa8-9805f1628bb7