Get rid of remaining Booleans-as-encoding-arguments in proto_tree_add_item() calls.
Change the "idl-light"-to-dissector generator to use ENC_NA for
structures and ENC_LITTLE_ENDIAN for bitmasks (are they always
little-endian, or should they check the data representation?).  Also
change it to properly set the display format for FT_ABSOLUTE_TIME fields
(it now specifies whether to show local time, UTC, or UTC with a
1-origin day-of-year).

Change one routine in the template for the DRSUAPI dissector to use
ENC_UTF_16|ENC_LITTLE_ENDIAN with proto_tree_add_item() (the field it
dissects is known to be a string, so there's no need to check its data
type).

Regenerate the BUDB, BUTC, and DRSUAPI dissectors.  (The generated BUTC
dissector handles pointers differently - either the old dissector was
wrong, or the generator is wrong, or the IDL for the protocol is wrong,
or the difference doesn't actually affect dissection.)

The generator doesn't currently handle hex values in enums correctly;
manually fix packet-dcerpc-drsuapi.h.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@42625 f5534014-38df-0310-8fa8-9805f1628bb7