Replace the File -> Export menu with separate:
	File -> Export Packet Dissections

	    (for the "print to file", "export as CSV", "export as C array",
	    "export as PSML", and "export as PDML" items)

	File-> Export Selected Packet Bytes

	File -> Export SSL Session Keys

	File -> Export Objects

	    (for exporting objects transferred over HTTP, DICOM, or SMB)

menu items.

The operations under Export really weren't that related - about all they
had in common was that they wrote to a file stuff other than packets
in a capture file format; the operations in the groups *under* Export
were related, so the groups are now menu items of their own.

This way, the File menu more immediately indicates what options of that
sort are available.

It also means that the Export Packet Dissections item might make it
clearer that what you get from that is *NOT* something that can just be
read back into Wireshark, as at least one user who asked "how do I get
my capture back from this?" on ask.wireshark.com thought.  If that
doesn't suffice, perhaps renaming it to "Export Dissected Packets" would
help; if *that* doesn't suffice, perhaps Kevin Cullimore's suggestion
that it say "Report" rather than "Export" will do the trick:

	From: Kevin Cullimore <kcullimo@runbox.com>
	Subject: [Wireshark-users] Re: Should the "export as text" item be in an "Export Human-readable..." item in the File menu?
	Date: May 19, 2012 8:31:23 PM PDT
	To: wireshark-users <wireshark-users@wireshark.org>

	Would classifying the asymmetric export (ones that lack a
	corresponding "import" action) formats as "reports" help clear
	up the original ambiguity/misunderstanding? It seems that most
	of the gui-based network tools I'm forced to periodically
	interact with rely upon that term with at least some success.

(Or perhaps some other verb would be right in some cases, e.g. "Save SSL
Session Keys".)

This also sets a pattern for another upcoming change - splitting "Save
As" into "Save As", which always saves every packet and makes the new
file the current file, and "{Verb} Specified Packets", which lets you
specify which packets to save and does *not* make the new file the
current file.  That'd simplify the code a bit, and might clear up the
new only-in-the-trunk issue in bug 6640 - having "Save As" default to
saving displayed packets currently means that it acts more like the
latter of those functions.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@42778 f5534014-38df-0310-8fa8-9805f1628bb7