vwr_read_rec_header()'s return value is used only to distinguish success (read a record header) from failure (got an EOF or an error).  Make it just return a Boolean.
If it fails in vwr_read(), don't overwrite *err_info (yes,
vwr_read_rec_header() might have set *err_info, so don't lose - and
leak! - the value it returned) - trust vwr_read_rec_header(), or the
routines it calls, to have set it.  (If there's a code path where that
doesn't happen, that code path needs to be fixed; the setting of
*err_info in vwr_read() should *not* be restored.)

Thanks to Evan Huus for finding a useless variable with cppcheck, and
reporting it in bug 7295, provoking me to look at this.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@42865 f5534014-38df-0310-8fa8-9805f1628bb7