Update some links (the "[Ff]amily[Ii]d=" links just redirected to the "id=" links when I last tried them).
There  are apparently both x86 and x86 versions of the VC++ 2005
distributables and, at least for the non-SP1 versions, there's also an
IA64 version.  List them.

If the URLs have the country and language in them, the displaylang= item
doesn't seem to work (tried with both "displaylang=en" and
"displaylang=de" with both "en-us" and "de-de").  Remove it.

Fix some typoes.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43026 f5534014-38df-0310-8fa8-9805f1628bb7