Add a preference to that when the window scaling options are not visible, we can still configure a scaling factor and therefore get better TCP sequence analysis and see better stream graphs.
A single preference is used for any/all streams for which the true value
isn't known.  I tried to make it obvious when showing the calculated
window size that it came fromt he preference setting.  The default value
for this preference is off, i.e. it won't change existing behaviour.

This was discussed a little at Sharkfest and raised on the developers
list last week.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@43686 f5534014-38df-0310-8fa8-9805f1628bb7