From Masashi Honma via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7691 TKIP dissection : wrong IS_TKIP macro
In [1] "11.4.2.2 TKIP MPDU formats", we could see below sentence.
"WEPSeed[1] is not used to construct the TSC, but is set to (TSC1 | 0x20) &
0x7f."

But the IS_TKIP macro only checks (WEPSeed[1] & 0x20).
So sometimes IS_TKIP macro mis-dissects a CCMP frame as a TKIP frame.

This patch changes IS_TKIP macro to do more better check.

[1] IEEE Std 802.11.-2012

#BACKPORT(1.8, 1.6)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@44790 f5534014-38df-0310-8fa8-9805f1628bb7