Just append the ": " to the generic NACK BLP item, don't replace its text.
Don't put empty items into the bitfield tree.

Show the BLP field as hex, as it's a bitfield.

(What should *really* happen is that it should display each 1 bit in the
usual fashion, with the ".... ...1 .... ...." display, but have that bit
be an *integral* field with the appropriate PID for the NACKed packet -
i.e., display hf_rtcp_rtpfb_nack_pid, with the value
rtcp_rtpfb_nack_pid + i + 1, and with the bitfield stuff in the
displayed representation.)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@45155 f5534014-38df-0310-8fa8-9805f1628bb7