For VS 2005, also mention the IA-64 vcredist the first time we mention the versions of vcredist, just in case anybody is building for Itanium. The second time we mention them, mention all three, not just the 32-bit x86 version.  Rewrite the sentence for that a bit - it's not as if MSVC uses particular deployment methods, those were chosen by the Wireshark development team.
For VS 2008, note that Microsoft mentions all three redistribution
mechanisms, and link to the "Choosing a Deployment Method" page for VS
2008 (which looks as if Microsoft didn't do a good job of editing it
after copying-and-pasting the 2005 version, or didn't set up ASP or
whatever it is that supplies VS-specific versions of the pages
correctly).

For VS 2010, note that Microsoft mentions all three redistribution
mechanisms, link to the "Choosing a Deployment Method" page for VS 2010,
and note that it recommends the method we chose and gives it as as the
first method.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@45302 f5534014-38df-0310-8fa8-9805f1628bb7