Add a PacketRangeGroupBox widget. Use it to implement "Export Selected Packets". Not yet tested on Windows. "Ignore Packet" hasn't been implemented so we can't test that either.
Create a SyntaxLineEdit widget from the QLineEdit code in
DisplayFilterEdit. Use it in the file import and export dialogs and the
PacketRangeGroupBox widget. This lets us provide instant feedback
instead of popping up an error dialog.

Expand the Tango color list based on
http://emilis.info/other/extended_tango .

Rearrange QtShark.pro to (hopefully) work better with Qt Creator.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@45405 f5534014-38df-0310-8fa8-9805f1628bb7