Fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7858
Use a full 32-bit literal instead of just a 16-bit one. Fixes case where
the value we're &-ing with just slips over 2^16, making us get stuck
in an infinite loop.

I'm not sure this matches the iscsi spec anymore, the comment in the code
about padding bytes is ambiguous as to whether they're leading or trailing.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@45524 f5534014-38df-0310-8fa8-9805f1628bb7