*Always* create a new data source when we decompress an IPv6 header.
Label the data sources for them as "6LoWPAN xxx", where "xxx" is the
type of compression, and label the data sources for reassembled
fragments as "Reassembled 6LoWPAN".

Fix the capitalization of "6LoWPAN".

Note that if reassembly fails, continuing dissection is not the right
thing to do, at least not if it failed because we don't *yet* have all
the fragments.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@45527 f5534014-38df-0310-8fa8-9805f1628bb7